backend/requirements.txt
Django==4.1.13
mongoengine==0.29.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
pymongo==4.5.0
python-decouple==3.8
Pillow==10.0.0
pyjwt==2.8.0
sendgrid==6.11.0
requests==2.31.0
dnspython==2.4.2
bcrypt==4.0.1